refactored drinkservice to pubservice and pubcontroller and added buy method

+++ pub/src/main/java/com/pub/repositories/DrinkRepository.java
@@ -4,5 +4,5 @@ import com.pub.models.Drink;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DrinkRepository extends JpaRepository<Drink, Integer> {
+public interface DrinkRepository extends JpaRepository<Drink, Long> {
 }
